+++ gradle.properties
@@ -1,8 +1,9 @@
-version=4.0.0-SNAPSHOT
-grailsVersion=4.0.11
-gorm.version=7.0.8.RELEASE
-gormHibernate5Version=7.0.5
-hibernateVersion=5.4.18.Final
+version=5.0.0-SNAPSHOT
+grailsVersion=5.3.5
+gorm.version=7.3.4
+gormHibernate5Version=7.3.1
+grailsGradlePluginVersion=5.3.1
+groovyVersion=3.0.11
 org.gradle.daemon=true
 org.gradle.parallel=true
 org.gradle.jvmargs=-Dfile.encoding=UTF-8 -Xmx1024M
