+++ grails-app/init/grails/plugins/taggable/Application.groovy
@@ -1,8 +1,11 @@
 package grails.plugins.taggable
 
+
 import grails.boot.GrailsApp
 import grails.boot.config.GrailsAutoConfiguration
+import grails.plugins.metadata.PluginSource
 
+@PluginSource
 class Application extends GrailsAutoConfiguration {
     static void main(String[] args) {
         GrailsApp.run(Application, args)
